Revise scan logic
Also scan contents of /System/Volumes/Data, but scan it last.
Ignore duplicates due to firmlinks by applying hardlink check to
folders.

This way files appear in /System/Volumes when the scan excludes
the main locations of these folders.
It should also make the scan more robust, as it does not rely on
hard-coded assumptions on folders that can be skipped.

The drawback is that scanning is slightly slower and requires
more temporary storage, due to the hard-link check for folders.